+++ wizard/__init__.py
@@ -1,2 +1,0 @@
-# Wizard Module
-from .wizard_utils import *